[Refac] seperated MiniCTA from ButtonTag
